feat(drand): Add League of Entropy distributed randomness beacon integration
Implements unified seeding interface with 6 strategies via gay_seed!:

1. Known (Deterministic): gay_seed!(42), gay_seed!("phrase")
2. Random, Known (Replay): gay_seed!(:record, round_id)
3. Random, Will Be Known: gay_seed!(:record) - returns DrandRound proof
4. Random, Unknown: gay_seed!(:new) - drand quicknet, no proof
5. Random, Unknown, Local: gay_seed!(:offline) - system entropy
6. Content-Addressable: gay_seed!(bytes) - SHA-256 of file/data

Also syncs src/ with upstream/gay to ensure all source files exist.

+++ src/exo_mlx.jl
@@ -321,7 +321,7 @@ function inject_spi_colors(hidden_states::Matrix{Float32}, layer::Int;
                               (UInt64(d) * 0xc4ceb9fe1a85ec53)
             r, _, _ = hash_color(h, UInt64(t))
             # Tiny additive color (doesn't affect model output significantly)
-            colored[t, d] += r * 1e-7f0
+            colored[t, d] += r * Float32(1e-7)
         end
     end
